# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Multi-Tenant Express API",
    "description": "Multi-tenant REST API with Organization → Project → Resource hierarchy, RBAC permissions, and invitation system.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "security": [
    {
      "AccessToken": []
    }
  ],
  "paths": {
...
    "/api/invitations/{invitation_id}/accept": {
      "post": {
        "tags": ["Invitations"],
        "summary": "Accept invitation",
        "description": "Accept a pending invitation to join an organization or project.",
        "operationId": "acceptInvitation",
        "security": [{ "AccessToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/InvitationIdParam" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invitation accepted successfully",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "message": "Invitation accepted",
                  "data": null
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/invitations/{invitation_id}/decline": {
      "post": {
        "tags": ["Invitations"],
        "summary": "Decline invitation",
        "description": "Decline a pending invitation.",
        "operationId": "declineInvitation",
        "security": [{ "AccessToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/InvitationIdParam" }],
        "responses": {
          "200": {
            "description": "Invitation declined successfully",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "message": "Invitation declined",
                  "data": null
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/orgs/{org_id}/invitations/{invitation_id}": {
      "delete": {
        "tags": ["Invitations"],
        "summary": "Revoke invitation",
        "description": "Revoke a pending invitation. Only users with appropriate permissions can revoke invitations.",
        "operationId": "revokeInvitation",
        "security": [{ "AccessToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/OrgIdParam" },
          { "$ref": "#/components/parameters/InvitationIdParam" }
        ],
        "responses": {
          "200": {
            "description": "Invitation revoked successfully",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "message": "Invitation revoked",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "JWT access token (httpOnly cookie set by server)"
      },
      "RefreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "JWT refresh token (httpOnly cookie set by server)"
      }
    },
    "headers": {
      "X-Request-Id": {
        "description": "Correlation ID for request tracing. Accepts a valid UUID or generates one.",
        "schema": {
          "type": "string",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        }
      }
    },
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          }
        },
        "required": ["message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"]
      },
...
      "AcceptInvitationRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 64,
            "maxLength": 64,
            "description": "Invitation token received by the invitee (64-character hex string)"
          }
        },
        "required": ["token"]
      }
    },
    "parameters": {
      "OrgIdParam": {
        "name": "org_id",
        "in": "path",
        "required": true,
        "description": "UUID of the organization",
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "example": "550e8400-e29b-41d4-a716-446655440000"
      },
...
      "InvitationIdParam": {
        "name": "invitation_id",
        "in": "path",
        "required": true,
        "description": "UUID of the invitation",
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "example": "990e8400-e29b-41d4-a716-446655440000"
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request data",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validation": {
                "summary": "Validation error",
                "value": {
                  "message": "\"name\" is required"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid credentials",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "noToken": {
                "summary": "No token provided",
                "value": {
                  "message": "No token provided"
                }
              },
              "invalid": {
                "summary": "Invalid token",
                "value": {
                  "message": "Invalid token"
                }
              },
              "expired": {
                "summary": "Token expired",
                "value": {
                  "message": "Token expired"
                }
              },
              "credentials": {
                "summary": "Invalid credentials",
                "value": {
                  "message": "invalid credentials"
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "You do not have permission to perform this action"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "resource": {
                "summary": "Resource not found",
                "value": {
                  "message": "Resource not found"
                }
              },
              "route": {
                "summary": "Route not found",
                "value": {
                  "message": "Not Found"
                }
              }
            }
          }
        }
      }
    }
  }
}